hyperlight_guest_bin: move exposed virtual memory operations to arch-specific module
This changes hyperlight_guest_bin::paging to be a reexport of the
interface of an architecture-specific implementation, much like
hyperlight_common::vmem.
